Removed the lemmata.sort_key column.
The intended use for this column was for specifying the sort order of
lemmata in order to emulate appropriate dictionary order wherever this
deviates from lexical string sorting. In practice it has not been used,
e.g. it is not used when sorting lemmata for end-user presentation. In
the PROIEL data set only the Gothic Bible has data in this column (with
page and column numbers from Streitberg's dictionary) and this
information is duplicated in lemmata.foreign_keys.