Removed the lemmata.full_gloss column.
The intended use for this column was for specifying the a longer gloss
of the lemma, in contrast to short_gloss, which was intended for a short
gloss that could be presented as an interlinear gloss. Only support for
short_gloss was ever implemented in the application, and this column is
entirely unused in the PROIEL data set.