Added sources.additional_metadata and removed TEI header from database.
This adds a new column called additional_metadata to the sources table.
The new column contains a serialised hash with metadata on top of the
metadata already represented by columns in the sources table. As it
includes fields that offer more details about the printed edition being
used, the edition column is superseded and removed from the sources
table. The tei_header column is also removed since it is now possible to
generate a TEI header from the metadata in the database.

The new fields can be edited in sources#edit and are shown in
sources#show.

A new version 2.0 of the PROIEL XML export format adds support for the
new fields. PROIEL XML 1.0 is still supported for imports but not for
exports. On import of PROIEL XML 1.0 the tei_header and edition elements
are ignored.