Allow a lemma's variant number to be any integer.
This removes a restriction enforced by MorphFeatures.new that the
variant number of a lemma should be a positive integer unless null.

The restriction was originally imposed to enforce a convention since the
end-user form of a lemma, ("cado#1", "cado#2", etc.) is easier to read
when variant numbers are positive.

Since the restriction is not enforced elsewhere the test in
MorphFeatures.new can trigger an ArgumentError exception that is not
handled properly by calling code.

As zero or negative variant numbers do not affect data integrity it is
preferable to remove the restriction rather than improve enforcement
(and thus error handling). Convention should instead be enforced by
reviewers and annotators, if desired.