Handle sorting by location in token searches manually.
This overrides the sort fields passed to ransack when the user specifies
sorting by location. This is slightly tricky to get right since so many
levels of objects are involved, but overriding the sorts attribute of the
search object on the basis of the params hash appears to produce acceptable
results.