Removed plugin system and updated external link mappers.
The plugin system (which allowed graph visualizers, sentence exporters
and external link mappers to be independent plugins) was never really
used much and simply introduces an extra layer of indirection that has
to be maintained. This patch removes it.

The external link mappers have also been updated: The one for bibelen.no
is no longer usable since the URLs are unpredictable. The biblos site
appears to have been renamed biblehub.com.