Write an ID mapping file if importing PROIEL XML with existing IDs.
This adds functionality to the PROIEL XML importer allowing the user
to save a mapping between ID attributes on objects in an imported PROIEL
XML file and new, generated ID attributes on objects in the PROIEL
database. The ID mapping is saved as a comma-separated file with a
user-specified name.