Fix: export PROIELXML with PRO tokens in correct order.
Until now PRO tokens (those with empty_token_sort set to 'P') have been
included in the PROIELXML export as part of the export of
information-structural annotation, but as they lack ordering they have
been positioned after all other tokens. This does not just look odd, it
also leads to confusing results in tools like TigerSearch since the user
cannot assume that PROs are adjacent to their heads when constructing
queries.

The presentation code in the application already repositions PROs next
to their heads so the logical thing to do is to apply the same ordering
principles in the export.

The patch changes the following:

* Apply the same ordering principles for PRO tokens in presentation code
  and PROIELXML export.
* Change the Sentence#tokens_with_dependents_and_info_structure
  association to a function tokens_with_deps_and_is and merge it with
  with_prodrops_in_place.
* Delegate special rendering of PRO tokens to the caller of
  Sentence#tokens_with_deps_and_is.