Change default web server to unicorn.
This replaces thin with unicorn and expands the instructions in
README.md to explain an alternative using Rails' own default server.

thin has been a source of problems: On our main server EventMachine
would suddenly start looping continuously, constantly polling and taking
up 100% of one CPU. This would happen consistently after 12-24 hours and
continue until restarted. Tracking down the source of the error proved
difficult. Using unicorn instead has eliminated the issue and as it is a
drop-in replacement for this application, we may as well use it as a
default from now on.