Move statistics charts to source views.
The global statistics charts that can be reached from the navbar cease
to be meaningful as the treebank grows. On their current form the
granularity of the charts showing contributors are in particular not
meaningful. There is code for showing per-source statistics, but is not
reachable. This patch removes the global statistics,enables
per-source statistics instead and eliminates resulting dead code.