Export texts by default to the current directory.
proiel:text:export now exports to the current directory unless
EXPORT_DIRECTORY has been set to something else. Also the directory will
not be created (since it makes little sense to insist on creating the
directory '.', for example, if that is what the caller specifies).

This eliminates the final use of the public/exports directory.