Add bin/proiel-webapp for maintenance tasks
This change removes all the rake the tasks in the proiel namespace that
addressed importing and exporting various pieces of data and replaces
them with a command-line utility (proiel-webapp) that generally performs
the same tasks.

The export/import formats were in most cases already comma-separated
files, and these have been standardised so that all have proper headers
and only contain data that can be round-tripped. This breaks backwards
compatibility in practically all cases and means that some tasks that
could be performed with a single task in the past now require several
tasks and an external script to combine them.

The following tasks were removed altogether:

* proiel:morphology:reassign: instead export using `export morphology`,
alter exported file, and re-import using `import morphology`

* proiel:morphology:force_manual_tags (broken): instead prepare a CSV
file and import using `import morphology`

* proiel:history:prune:attribute (export using `export history`, remove
attribute from exported file, clear using `delete history`, and import
using `import history`)

All new tasks roundtrip except for `history` (which only supports
`delete`) and `text` (which does not have `delete`). There is not yet
any guarantee that roundtripping does not lead to data loss as there may
be attributes on objects that shhould be exported on export but have
been overlooked.